pyproject.toml
[project]
name = "kro"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "anthropic>=0.64.0",
    "dotenv>=0.9.9",
    "mkdocs-material[recommended]>=9.6.16",
    "pymupdf4llm>=0.0.27",
    "pyyaml>=6.0.2",
    "ufal-morphodita>=1.11.2.1",
]
